requirements.txt
pyrogram>=2.0.106
python-dotenv>=1.0.1
tgcrypto>=1.2.5
uvloop>=0.19.0
supabase